CHG: remove the hippo kernels from s4 module
